resolve PR issues - add documentation - split initialization - isolate network layer for Iroha - isolate IrohaCryptoProvider
