+++ build.gradle
@@ -33,11 +33,6 @@ dependencies {
     compile 'io.grpc:grpc-netty:1.8.0'
     compile 'io.grpc:grpc-protobuf:1.8.0'
     compile 'io.grpc:grpc-stub:1.8.0'
-    if (JavaVersion.current().isJava9Compatible()) {
-        // Workaround for @javax.annotation.Generated
-        // see: https://github.com/grpc/grpc-java/issues/3633
-        compile 'javax.annotation:javax.annotation-api:1.3.1'
-    }
 
     // ------------------| Src dependencies |------------------
