+++ src/main/kotlin/main/Configs.kt
@@ -6,20 +6,20 @@ package main
 object Configs {
 
     /** Port for refund REST API */
-    var refundPort = 8080
+    val refundPort = 8080
 
     /** Path to public key */
-    var pubkeyPath = "example/admin@test.pub"
+    val pubkeyPath = "example/admin@test.pub"
 
     /** Path to private key */
-    var privkeyPath = "example/admin@test.priv"
+    val privkeyPath = "example/admin@test.priv"
 
     /** Iroha account for creator txs */
-    var irohaCreator = "admin@test"
+    val irohaCreator = "admin@test"
 
     /** Iroha peer hsotname */
-    var irohaHostname = "localhost"
+    val irohaHostname = "localhost"
 
     /** Iroha peer hsotname */
-    var irohaPort = 50051
+    val irohaPort = 50051
 }
