+++ src/main/kotlin/sideChain/iroha/consumer/IrohaCryptoProvider.kt
@@ -1,0 +1,11 @@
+package sideChain.iroha.consumer
+
+import Keypair
+
+/**
+ * Provides with Iroha keypair
+ */
+interface IrohaCryptoProvider {
+    /** Iroha public and private key */
+    val keypair: Keypair
+}
