+++ src/main/kotlin/sideChain/iroha/consumer/IrohaNetwork.kt
@@ -1,0 +1,14 @@
+package sideChain.iroha.consumer
+
+import iroha.protocol.BlockOuterClass
+
+/**
+ * Interface for network layer of Iroha chain
+ */
+interface IrohaNetwork {
+    /**
+     * Send transaction to Iroha
+     * @param protoTx protobuf representation of transaction
+     */
+    fun send(protoTx: BlockOuterClass.Transaction)
+}
