+++ src/test/notary/IrohaCommandTest.kt
